Update TravelOptionsMapWindow.cs
This is a proposed change to implement the following idea: https://forums.dfworkshop.net/viewtopic.php?f=12&t=5117.

I have added a message receiver within my Hidden Map Locations mod, in order to populate the 'discoveredMapSummaries' and  'revealedLocationTypes' properties within this class.

The 'discoveredMapSummaries' is a collection stored within my mod to determine whether the player has discovered the location or not. The 'revealedLocationTypes' is a collection of location types that the player has decided that they want revealed by default, configured in their mod settings (eg. Cities).

So when Travel Options draws the discovered map locations it will now check if the Hidden Map Locations mod is enabled, and if so use these 2 collections to determine whether or not to draw the map pixel. Otherwise it will function as it did before.

I didn't want to put any message senders within the 'checkLocationDiscovered' method, as it would be called many times within the loop.